Migratating the diakonia project documentation published files from their own repository to a specific folder of the master branch of the main repo for ouh-churchill.
Can't publish into the same docs folder that the sources were originally in, so have renamed `docs` to `docs-source`, and generated a new blank `docs` folder which unfortunately results in a domain/projectname/ url, rather than hosted at the root of the domain!

So now we're trying to find a way to handle that in GitHub workarounds, since trying to do a Travis process that publishes to a secondary repository from within it's own is going to be very challenging.